MCPs on Databricks celebal/ai-data-discovery-assistant/requirements.txt
mcp>=1.26.0
fastmcp>=3.1.1
requests
python-dotenv
pydantic
uvicorn
